# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/shell-terminals:
    get:
      operationId: listShellTerminals
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListShellTerminalsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List the standalone shell terminals owned by the current app run
      tags:
      - shellTerminals
    post:
      operationId: openShellTerminal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenShellTerminalRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellTerminalEnvelope'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Open a standalone shell terminal
      tags:
      - shellTerminals
  /api/v1/shell-terminals/{handleId}:
    delete:
      operationId: closeShellTerminal
      parameters:
      - description: Shell terminal runtime handle identifier.
        in: path
        name: handleId
        required: true
        schema:
          description: Shell terminal runtime handle identifier.
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Close a standalone shell terminal and destroy its PTY
      tags:
      - shellTerminals
    patch:
      operationId: renameShellTerminal
      parameters:
      - description: Shell terminal runtime handle identifier.
        in: path
        name: handleId
        required: true
        schema:
          description: Shell terminal runtime handle identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShellTerminalRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellTerminalEnvelope'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Rename a standalone shell terminal tab
      tags:
      - shellTerminals
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    ListShellTerminalsResponse:
      properties:
        shellTerminals:
          items:
            $ref: '#/components/schemas/ShellTerminalResponse'
          type: array
      required:
      - shellTerminals
      type: object
...
    OpenShellTerminalRequest:
      properties:
        projectId:
          description: Project whose root the shell starts in. Omitted opens the shell
            in the daemon data dir.
          type: string
        sessionId:
          description: Agent session the shell is scoped to, so it appears only in
            that session's tab strip. Omitted makes it a standalone shell.
          type: string
        shell:
          description: 'Windows shell selector: auto, git-bash, pwsh, powershell,
            cmd, or a custom executable path. Ignored on macOS and Linux.'
          type: string
      type: object
...
    ShellTerminalEnvelope:
      properties:
        shellTerminal:
          $ref: '#/components/schemas/ShellTerminalResponse'
      required:
      - shellTerminal
      type: object
    ShellTerminalResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        handleId:
          type: string
        projectId:
          type: string
        sessionId:
          type: string
        title:
          type: string
        workingDir:
          type: string
      required:
      - handleId
      - workingDir
      - title
      - createdAt
      type: object
...
    UpdateShellTerminalRequest:
      properties:
        title:
          description: New tab title for the shell terminal. Trimmed; must be non-empty.
          type: string
      required:
      - title
      type: object
...